INDY-1876: update tests and fix in on_catch_up_finished
