+++ plenum/test/replica/stashing/test_stash_future_view.py
@@ -70,95 +70,3 @@ def test_process_three_phase_msg_and_stashed_future_view(txnPoolNodeSet, looper,
 
         looper.run(eventually(chk))
         waitNodeDataEquality(looper, slow_node, *fast_nodes)
-
-
-def test_unstash_three_phase_msg_after_catchup_in_view_change(txnPoolNodeSet, looper, tconf,
-                                                         sdk_pool_handle,
-                                                         sdk_wallet_steward):
-    """
-    1. Delay Commit on Node4
-    2. Order 1 req
-    3. Delay Commit on all nodes
-    4. Order 1 req
-    5. Delay CatchupRep on Node4
-    6. Delay Ledger Status on Nodes1-3
-    7. Start View change on all nodes
-    8. Wait until Node4 got 3 stashed CatchupReps
-    9. reset delaying of Commits on all Nodes
-    10. Reset Ledger Status on Nodes1-3
-    11. check that 3 nodes finished VC while Node4 is syncing and not finished
-    12. reset CatchupRep on Node
-    13. check that Node4 finished VC, and there was just 1 round of cacth-up (edited)
-    """
-    slow_node = txnPoolNodeSet[-1]
-    fast_nodes = txnPoolNodeSet[:-1]
-    view_no = txnPoolNodeSet[0].viewNo
-    old_stashed = slow_node.master_replica.stasher.num_stashed_future_view
-    last_ordered = txnPoolNodeSet[0].master_replica.last_ordered_3pc
-
-    with delay_rules([n.nodeIbStasher for n in txnPoolNodeSet],
-                     msg_rep_delay(types_to_delay=[PREPREPARE, PREPARE, COMMIT])):
-
-        # Delay Commit messages for slow_node.
-        slow_node.nodeIbStasher.delay(cDelay(sys.maxsize))
-        sdk_send_random_and_check(looper, txnPoolNodeSet, sdk_pool_handle,
-                                  sdk_wallet_steward, 1)
-
-        # Delay Commit messages for fast_nodes.
-        for n in fast_nodes:
-            n.nodeIbStasher.delay(cDelay(sys.maxsize))
-
-        request2 = sdk_send_random_request(looper, sdk_pool_handle, sdk_wallet_steward)
-
-        def check_commits(commit_key):
-            for n in txnPoolNodeSet:
-                assert all(len(r.commits[commit_key].voters) == 1 for r in n.replicas.values())
-
-        looper.run(eventually(check_commits,
-                              (view_no, last_ordered[1] + 2)))
-
-        # Delay CatchupRep messages for the slow_node.
-        with delay_rules([slow_node.nodeIbStasher], cr_delay()):
-            with delay_rules([n.nodeIbStasher for n in fast_nodes], vcd_delay()):
-                with delay_rules([n.nodeIbStasher for n in fast_nodes],
-                                 msg_rep_delay(types_to_delay=[LEDGER_STATUS])):
-
-                    for n in txnPoolNodeSet:
-                        n.view_changer.on_master_degradation()
-                    looper.run(eventually(lambda: assertExp(slow_node.mode == Mode.syncing)))
-
-                    # Reset delay Commit messages for all nodes.
-                    for n in txnPoolNodeSet:
-                        n.nodeIbStasher.reset_delays_and_process_delayeds(COMMIT)
-
-                    assert slow_node.view_change_in_progress
-                    assert slow_node.mode == Mode.syncing
-                    looper.run(eventually(_check_nodes_stashed,
-                                          fast_nodes,
-                                          old_stashed,
-                                          len(txnPoolNodeSet) - 1))
-                    looper.run(eventually(_check_nodes_stashed,
-                                          [slow_node],
-                                          old_stashed,
-                                          (len(txnPoolNodeSet) - 1) * 2))
-
-            waitForViewChange(looper, fast_nodes, expectedViewNo=view_no + 1,
-                              customTimeout=2 * tconf.VIEW_CHANGE_TIMEOUT)
-            ensureElectionsDone(looper=looper,
-                                nodes=fast_nodes,
-                                instances_list=range(fast_nodes[0].requiredNumberOfInstances),
-                                customTimeout=2 * tconf.VIEW_CHANGE_TIMEOUT)
-        sdk_get_and_check_replies(looper, [request2])
-        waitForViewChange(looper, [slow_node], expectedViewNo=view_no + 1,
-                          customTimeout=2 * tconf.VIEW_CHANGE_TIMEOUT)
-        ensureElectionsDone(looper=looper,
-                            nodes=txnPoolNodeSet)
-        _check_nodes_stashed(fast_nodes, old_stashed, 0)
-        assert all(n.master_replica.last_ordered_3pc == (last_ordered[0],
-                                                         last_ordered[1] + 2)
-                   for n in txnPoolNodeSet)
-
-
-def _check_nodes_stashed(nodes, old_stashed, new_stashed):
-    for n in nodes:
-        assert n.master_replica.stasher.num_stashed_catchup == old_stashed + new_stashed
